# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Symbiota API",
        "contact": {
            "name": "Symbiota Support Hub Team",
            "email": "symbiota@asu.edu"
        },
        "version": "2.0"
    },
    "servers": [
        {
            "url": "../"
        }
    ],
    "paths": {
...
        "/api/v2/exsiccata": {
            "get": {
                "tags": [
                    "Exsiccata"
                ],
                "operationId": "/api/v2/exsiccata",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Controls the number of results in the page.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Determines the starting point for the search results. A limit of 100 and offset of 200, will display 100 records starting the 200th record.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns list of inventories registered within system",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. "
                    }
                }
            }
        },
        "/api/v2/exsiccata/{identifier}": {
            "get": {
                "tags": [
                    "Exsiccata"
                ],
                "operationId": "/api/v2/exsiccata/{identifier}",
                "parameters": [
                    {
                        "name": "identifier",
                        "in": "path",
                        "description": "Exsiccata ID or record ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns exsiccata record with matching identifier",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. Valid Exsiccata identifier required"
                    },
                    "404": {
                        "description": "Record not found"
                    }
                }
            }
        },
        "/api/v2/exsiccata/{identifier}/number": {
            "get": {
                "tags": [
                    "Exsiccata"
                ],
                "operationId": "/api/v2/exsiccata/identifier/number",
                "parameters": [
                    {
                        "name": "identifier",
                        "in": "path",
                        "description": "Identifier (ometid (PK) - currently does not accommodate recordID) associated with target exsiccata title",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Controls the number of results in the page.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Determines the starting point for the search results. A limit of 100 and offset of 200, will display 100 records starting the 200th record.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns all exsiccata numbers associated with a single exsiccati title corresponding to matching identifier",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. Exsiccata identifier is required."
                    },
                    "404": {
                        "description": "Record not found"
                    }
                }
            }
        },
        "/api/v2/exsiccata/{identifier}/number/{numberIdentifier}/": {
            "get": {
                "tags": [
                    "Exsiccata"
                ],
                "operationId": "/api/v2/exsiccata/identifier/number/{numberIdentifier}/",
                "parameters": [
                    {
                        "name": "identifier",
                        "in": "path",
                        "description": "Identifier (ometid (PK)) associated with target exsiccata number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "numberIdentifier",
                        "in": "path",
                        "description": "Identifier (omenid or recordID in the schema) associated with target exsiccata number",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Controls the number of results in the page.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Determines the starting point for the search results. A limit of 100 and offset of 200, will display 100 records starting the 200th record.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns all exsiccata numbers associated with a single exsiccati title corresponding to matching identifier",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: Bad request. Exsiccata identifier is required."
                    },
                    "404": {
                        "description": "Record not found"
                    }
                }
            }
        },
...
    },
...
}